@@ -19,32 +19,28 @@
   "dependencies": {
     "apollo-server-lambda": "1.2.0",
     "apollo-server-module-graphiql": "1.2.0",
-    "babel-polyfill": "^6.16.0",
-    "babel-runtime": "^6.25.0",
+    "babel-polyfill": "^6.26.0",
     "graphql": "^0.10.5",
     "graphql-playground-middleware": "^1.1.0",
     "graphql-tools": "2.7.2",
     "serverless-dynamodb-client": "^0.0.2"
   },
   "devDependencies": {
-    "babel-cli": "^6.16.0",
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.4.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
+    "aws-sdk": "^2.4.8",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
     "body-parser": "^1.15.2",
     "cors": "2.8.4",
     "env-cmd": "^5.0.0",
+    "faker": "^4.1.0",
     "imports-loader": "^0.7.1",
     "jest": "^21.2.1",
-    "serverless-offline": "^3.15.3",
-    "serverless-webpack": "^2.2.0",
-    "webpack": "^2.2.0",
-    "webpack-node-externals": "^1.6.0",
-    "serverless-dynamodb-local": "^0.2.26",
     "jsonfile": "^4.0.0",
-    "faker": "^4.1.0",
-    "aws-sdk": "^2.4.8"
+    "serverless-dynamodb-local": "^0.2.26",
+    "serverless-offline": "^3.15.3",
+    "serverless-webpack": "^4.1.0",
+    "webpack": "^3.8.1",
+    "webpack-node-externals": "^1.6.0"
   }
 }
